requirements.txt
# ============================================================
#  VoC Risk Sentinel — requirements.txt
#  Python 3.11+ 권장 (개발 환경: 3.13)
# ============================================================

# ── LLM API ──────────────────────────────────────────────────
google-genai>=1.65.0

# ── 데이터 처리 ───────────────────────────────────────────────
pandas>=2.3.3
tqdm>=4.67.3

# ── 환경변수 관리 ─────────────────────────────────────────────
python-dotenv>=1.2.2

# ── 대시보드 ──────────────────────────────────────────────────
streamlit>=1.55.0
plotly>=6.6.0
